refactor y logout user

+++ Server/src/Handlers/specialty/deleteSpecialtyHandler.js
@@ -1,20 +1,20 @@
-const { Specialty } = require('../DB_connection');
-const deleteReg = require("../controllers/deleteReg");
-const showLog = require("../functions/showLog");
-const checkToken = require('../functions/checkToken');
+const { Specialty } = require('../../DB_connection');
+const deleteReg = require("../../controllers/deleteReg");
+const showLog = require("../../functions/showLog");
+const checkToken = require('../../functions/checkToken');
 
 const deleteSpecialtyHandler = async (req, res) => {
-  showLog(`deleteSpecialtyHandler`);
   try {
     const { id } = req.params;
     const { token } = req.query;
+    showLog(`deleteSpecialtyHandler (tkn ${token})`);
     // Verifico token. Sólo un superAdmin puede eliminar:
-    // if (!token) { throw Error("Token required"); }
-    // const checked = await checkToken(token);
-    // if (!checked.exist || checked.role !== "superAdmin") {
-    //     showLog(`Wrong token.`);
-    //     return res.status(401).send(`Unauthorized.`);
-    // }
+    if (!token) { throw Error("Se requiere token"); }
+    const checked = await checkToken(token);
+    if (!checked.exist || checked.role !== "superAdmin") {
+      showLog(`Wrong token.`);
+      return res.status(401).send(`Sin permiso.`);
+    }
     const resp = await deleteReg(Specialty, id, "Specialty");
     if (resp.deleted === 'ok') {
       showLog(`deleteSpecialtyHandler OK`);
